Se cambia prueba de api desde TiendaPage a CardShopComponent
