+++ src/app/components/card-shop/card-shop.component.spec.ts
@@ -1,28 +1,33 @@
-import { ComponentFixture, TestBed, waitForAsync } from '@angular/core/testing';
+import { ComponentFixture, TestBed, waitForAsync, fakeAsync, tick } from '@angular/core/testing';
 import { IonicModule } from '@ionic/angular';
-import { of } from 'rxjs';
 import { CardShopComponent } from './card-shop.component';
 import { MarvelApiService } from 'src/app/services/marvel-api.service';
+import { responseMock } from 'src/app/mocks/response.mock';
+import { of } from 'rxjs';
 
-
-let marvelApiServiceMock: any;
 describe('CardShopComponent', () => {
   let component: CardShopComponent;
   let fixture: ComponentFixture<CardShopComponent>;
+  let service: MarvelApiService;
 
   beforeEach(waitForAsync(() => {
-    marvelApiServiceMock = jasmine.createSpyObj('MarvelApiService', ['obtenerComics']);
+    let marvelApiServiceMock = {
+      obtenerComics: jasmine.createSpy('obtenerComics').and.returnValue(of(responseMock))
+    };
     TestBed.configureTestingModule({
       declarations: [ CardShopComponent ],
       imports: [IonicModule.forRoot()], 
       providers:[
-        {provide: MarvelApiService, useValue: marvelApiServiceMock}
-      ]
+        {
+          provide: MarvelApiService, 
+          useValue: marvelApiServiceMock}
+      ],
     }).compileComponents();
 
     fixture = TestBed.createComponent(CardShopComponent);
     component = fixture.componentInstance;
     fixture.detectChanges();
+    service = TestBed.inject(MarvelApiService);
   }));
 
   it('should create', () => {
@@ -45,5 +50,19 @@ describe('CardShopComponent', () => {
     component.onResize(mockEvent as unknown as Event);
     expect(component.setColumnSize).toHaveBeenCalledWith(800);
   });
+
+  it('Debería cargar los cómics desde localStorage si están disponibles', () => {
+    localStorage.setItem('comics', JSON.stringify(responseMock.data.results));
+    fixture.detectChanges();
+    expect(component.comics).toEqual(responseMock.data.results);
+  });
+
+  it('Debería obtener los cómics desde MarvelApiService si no hay cómics en localStorage', fakeAsync(() => {
+    localStorage.removeItem('comics');
+    component.ngOnInit();
+    tick();
+    expect(component.comics).toEqual(responseMock.data.results);
+    expect(service.obtenerComics).toHaveBeenCalled();
+  }));
 
 });
