+++ src/app/pages/leyendo/leyendo.page.ts
@@ -1,15 +1,14 @@
-import { Component, OnInit } from '@angular/core';
+import { Component } from '@angular/core';
 
 @Component({
   selector: 'app-leyendo',
   templateUrl: './leyendo.page.html',
   styleUrls: ['./leyendo.page.scss'],
 })
-export class LeyendoPage implements OnInit {
+export class LeyendoPage  {
 
   constructor() { }
 
-  ngOnInit() {
-  }
+ 
 
 }
